requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-filter==21.1
django-pure-admin==0.15
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
mysql-connector-python==8.0.24
mysqlclient==2.0.3
Pillow==8.4.0
protobuf==3.15.8
psycopg2==2.9.2
pycodestyle==2.7.0
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
